Move function to different file
